-----------------------------------------------------
fnspo and nspo options on gradual are now redundant.

-----------------------------------------------------

gradual does not apply the edge filter anymore

-----------------------------------------------------

Three new routines to detect of confirm scale-space extrema:
1) keypoints_find_3d_discrete_extrema_epsilon_sphere
2) keypoints_check_3d_discrete_extrema_epsilon_sphere
3) keypoints_confirm_extremum_present_in_ball

Details:

1) To look for extrema, each sample is compared to the samples inside points on
an annulus bounded by radii r and (r + dr)

2) The same criterion is applied at the nearest discrete sample to confirm the
existence of an extremum on that given point,

3) Another criterion. The mean, min or max value of samples inside a ball of
radius r1 is compared to the mean, min of max of samples inside an annulus
bounded by radii r2 and r3.
Three different criteria are coded, two of which are commented.

For these routines to work correcty, octaves are computed entirely (in a
gradual way) with
- sift_malloc_scalespace_dog_gradual_floatnspo_ENTIREOCTAVE

The corresponding binaries are
- gradual_check_extrema.cpp
- gradual_entireoctave.cpp

-----------------------------------------------------