New functionnalities to test and visualize extrema
Save scale-space file in binary files
- sift_write_scalespace_binary_file function
- sift_read_scalespace_binary_file function
with the option -bin_ss [label]  in extra_sift_cli.cpp

For a list of keypoints
- grab_ss_neighborhood program extracts the keypoint
surroundings and save the corresponding array in binary files.

For a surrouding stored in a binary file, 'check_extremum_in_cube.cpp' confirms
the presence of the extremum.